Hopefully fix failing test?

+++ ale/base/type_sensor.py
@@ -142,7 +142,7 @@ class Radar():
           ephemeris times split based on image lines
         """
         # 0.25 is the delta used by minirf, used as a default.
-        num_states = (self.ephemeris_stop_time - self.ephemeris_start_time)/0.25 + 1
+        num_states = int((self.ephemeris_stop_time - self.ephemeris_start_time)/0.25) + 1
         return np.linspace(self.ephemeris_start_time,  self.ephemeris_stop_time, num_states)
 
     @property
